+++ CHANGELOG.md
@@ -11,6 +11,12 @@ Changelog
 
 _Note: Gaps between patch versions are faulty, broken or test releases._
 
+## v3.64.0 (2021-11-16)
+
+#### :rocket: New Feature
+
+* Added a new method `unwrap` `core/promise/sync`
+
 ## v3.63.0 (2021-11-16)
 
 #### :rocket: New Feature